fix: indent special modes fallback
